add api-cards permissions

+++ api/views/card.py
@@ -6,33 +6,47 @@ from tasks.models import ListModel, BoardModel, CardModel, LabelModel, SubTaskMo
 from api.serializers.card import ListSerializer, LabelSerializer, SubCardSerializer, CardSerializer
 from accounts.models import User
 from rest_framework import status
+from permissions import IsCardMember, IsBoardMember, IsBoardOwner
 
 
 class MyCardsAPIView(APIView):
+    permission_classes = [IsCardMember,]
+
     def get(self, request, user_id):
         user = User.objects.get(id=user_id)
         cards = CardModel.objects.filter(cmembership__user=user)
+        self.check_object_permissions(request, cards)
         serializer = CardSerializer(cards, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 class CardsAPIView(APIView):
+    permission_classes = [IsBoardMember,]
+
     def get(self, request, board_id):
         query_set = CardModel.objects.filter(list__board_id=board_id)
+        board = get_object_or_404(BoardModel, id=board_id)
+        self.check_object_permissions(request, board)
         srz_data = CardSerializer(instance=query_set, many=True)
         return Response(data=srz_data.data, status=status.HTTP_200_OK)
 
 
 class CardAPIView(APIView):
+    permission_classes = [IsBoardMember,]
+
     def get(self, request, card_id):
         card = get_object_or_404(CardModel, id=card_id)
+        self.check_object_permissions(request, card.list.board)
         srz_data = CardSerializer(instance=card)
         return Response(data=srz_data.data, status=status.HTTP_200_OK)
 
 
 class CardCreateAPIView(APIView):
+    permission_classes = [IsBoardOwner,]
+
     def post(self, request, list_id):
         list_instance = get_object_or_404(ListModel, id=list_id)
+        self.check_object_permissions(request, list_instance.board)
         srz_data = CardSerializer(data=request.data)
         if srz_data.is_valid():
             srz_data.object.list = list_instance
@@ -42,8 +56,12 @@ class CardCreateAPIView(APIView):
 
 
 class CardUpdateAPIView(APIView):
+    permission_classes = [IsBoardOwner | IsCardMember]
+
     def put(self, request, card_id):
         card = get_object_or_404(CardModel, id=card_id)
+        self.check_object_permissions(request, card)
+        self.check_object_permissions(request, card.list.board)
         srz_data = CardSerializer(instance=card, data=request.POST, partial=True)
         if srz_data.is_valid():
             srz_data.save()
@@ -52,8 +70,11 @@ class CardUpdateAPIView(APIView):
 
 
 class CardDeleteAPIView(APIView):
+    permission_classes = [IsBoardOwner]
+
     def delete(self, request, card_id):
         card = get_object_or_404(CardModel, id=card_id)
+        self.check_object_permissions(request, card.list.board)
         card.delete()
         return Response({'message': 'card deleted'}, status=status.HTTP_200_OK)
